ci: use SeedImage for all CLI tests
