+++ .github/workflows/cli-k3s-obs_dev.yaml
@@ -38,7 +38,7 @@ jobs:
       cluster_name: cluster-k3s
       cluster_type: ${{ inputs.cluster_type }}
       destroy_runner: ${{ inputs.destroy_runner }}
-      iso_to_test: https://download.opensuse.org/repositories/isv:/Rancher:/Elemental:/Dev/containers/iso/elemental-teal.x86_64.iso
       k8s_version_to_provision: v1.26.8+k3s1
       operator_repo: oci://registry.opensuse.org/isv/rancher/elemental/dev/charts/rancher
+      os_to_test: dev
       rancher_version: ${{ inputs.rancher_version }}
