+++ .github/workflows/cli-k3s-os-upgrade.yaml
@@ -12,10 +12,6 @@ on:
         description: Destroy the auto-generated self-hosted runner
         default: true
         type: boolean
-      iso_to_test:
-        description: ISO to test
-        default: https://updates.suse.com/SUSE/Products/ElementalTeal/5.4/x86_64/iso/elemental-teal.x86_64-1.2.2-GM.iso
-        type: string
       operator_repo:
         description: Operator version to use for initial deployment
         default: oci://registry.opensuse.org/rancher
@@ -53,11 +49,11 @@ jobs:
       qase_run_id: ${{ inputs.qase_run_id }}
       cluster_name: cluster-k3s
       destroy_runner: ${{ inputs.destroy_runner }}
-      iso_to_test: ${{ inputs.iso_to_test }}
       k8s_version_to_provision: v1.26.8+k3s1
       node_number: 5
       operator_upgrade: oci://registry.opensuse.org/isv/rancher/elemental/${{ inputs.upgrade_os_channel }}/charts/rancher
       operator_repo: ${{ inputs.operator_repo }}
+      os_to_test: stable
       rancher_upgrade: ${{ inputs.rancher_upgrade }}
       rancher_version: ${{ inputs.rancher_version }}
       upgrade_image: registry.opensuse.org/isv/rancher/elemental/${{ inputs.upgrade_os_channel }}/containers/rancher/elemental-teal/${{ inputs.teal_version }}:latest
