+++ .github/workflows/cli-rke2-ibs_stable.yaml
@@ -39,8 +39,8 @@ jobs:
       cluster_name: cluster-rke2
       cluster_type: ${{ inputs.cluster_type }}
       destroy_runner: ${{ inputs.destroy_runner }}
-      iso_to_test: https://updates.suse.com/SUSE/Products/ElementalTeal/5.4/x86_64/iso/elemental-teal.x86_64-1.2.2-GM.iso
       k8s_version_to_provision: v1.26.8+rke2r1
       operator_repo: oci://registry.suse.com/rancher
+      os_to_test: stable
       rancher_version: ${{ inputs.rancher_version }}
       upstream_cluster_version: v1.26.8+rke2r1
