+++ .github/workflows/cli-rke2-obs_dev.yaml
@@ -39,8 +39,8 @@ jobs:
       cluster_name: cluster-rke2
       cluster_type: ${{ inputs.cluster_type }}
       destroy_runner: ${{ inputs.destroy_runner }}
-      iso_to_test: https://download.opensuse.org/repositories/isv:/Rancher:/Elemental:/Dev/containers/iso/elemental-teal.x86_64.iso
       k8s_version_to_provision: v1.26.8+rke2r1
       operator_repo: oci://registry.opensuse.org/isv/rancher/elemental/dev/charts/rancher
+      os_to_test: dev
       rancher_version: ${{ inputs.rancher_version }}
       upstream_cluster_version: v1.26.8+rke2r1
