+++ .github/workflows/cli-rke2-rancher_latest.yaml
@@ -1,9 +1,6 @@
 # This workflow calls the master E2E workflow with custom variables
 name: CLI-RKE2-Rancher_Latest
 
-# This worflow is scheduled *after* build-ci to be sure that we have
-# the lastest version built. The scheduling is also to avoid running
-# the worklow on each push on main.
 on:
   workflow_dispatch:
   schedule:
