+++ .github/workflows/cli-rke2-rancher_stable.yaml
@@ -3,13 +3,8 @@ name: CLI-RKE2-Rancher_Stable
 
 on:
   workflow_dispatch:
-  workflow_run:
-    workflows:
-      - build-ci
-    branches:
-      - main
-    types:
-      - completed
+  schedule:
+    - cron: '0 1 * * *'
 
 jobs:
   cli:
@@ -23,6 +18,4 @@ jobs:
       ca_type: private
       cluster_name: cluster-rke2
       k8s_version_to_provision: v1.26.8+rke2r1
-      start_condition: ${{ github.event.workflow_run.conclusion }}
-      workflow_download: ${{ github.event.workflow_run.workflow_id }}
       upstream_cluster_version: v1.26.8+rke2r1
