+++ tests/Makefile
@@ -83,6 +83,8 @@ e2e-install-app: deps
 	ginkgo --label-filter install-app -r -v ./e2e
 e2e-install-backup-restore: deps
 	ginkgo --label-filter install-backup-restore -r -v ./e2e
+e2e-iso-image: deps
+	ginkgo --label-filter iso-image -r -v ./e2e
 e2e-ui-rancher: deps
 	ginkgo --label-filter ui -r -v ./e2e
 e2e-uninstall-operator:
